[Refactor] cache
