+++ src/api/server.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"api/controller"
+	"core/cache"
 	"core/dao"
 	"core/logger"
 )
@@ -9,5 +10,6 @@ import (
 func main() {
 	logger.Init()
 	dao.Init()
+	cache.Init()
 	controller.Init()
 }
